+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -60,5 +60,5 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = dateutil.parser.isoparse("2025-03-04T22:54:39.742Z")
+    created: Optional[datetime] = dateutil.parser.isoparse("2025-03-05T18:46:01.581Z")
     r"""The date and time the resource was created"""