+++ packages/orq-rc/RELEASES.md
@@ -649,3 +649,13 @@ Based on:
 - [python v3.2.0-rc.27] packages/orq-rc
 ### Releases
 - [PyPI v3.2.0-rc.27] https://pypi.org/project/orq-ai-sdk/3.2.0-rc.27 - packages/orq-rc
+
+## 2025-03-05 18:47:08
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.509.1 (2.539.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.2.0-rc.28] packages/orq-rc
+### Releases
+- [PyPI v3.2.0-rc.28] https://pypi.org/project/orq-ai-sdk/3.2.0-rc.28 - packages/orq-rc